Fix FOV option type for Minecraft 1.21.11

+++ src/main/java/com/pixelforge/module/modules/utility/ZoomModule.java
@@ -5,10 +5,9 @@ import com.pixelforge.module.Module;
 import org.lwjgl.glfw.GLFW;
 
 public class ZoomModule extends Module {
-
     private boolean zooming;
-    private double previousFov = 70.0;
-    private final double zoomFov = 20.0;
+    private int previousFov = 70;
+    private final int zoomFov = 20;
 
     public ZoomModule() {
         super("Zoom", "OptiFine-style zoom (hold key)", Category.UTILITY);
@@ -19,10 +18,8 @@ public class ZoomModule extends Module {
     @Override
     public void onTick() {
         if (mc == null || mc.options == null) return;
-
         long window = mc.getWindow().getHandle();
         boolean keyDown = GLFW.glfwGetKey(window, getKeybind()) == GLFW.GLFW_PRESS;
-
         if (keyDown && !zooming) {
             previousFov = mc.options.getFov().getValue();
             mc.options.getFov().setValue(zoomFov);
